==> rmds/1.1_Fit_others.Rmd <==
---
title: "Fit other methods"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE)
library(magrittr)
library(ggplot2)
```

```{r setup2, include=FALSE}
# batchtools::makeRegistry(file.dir = "r_batchtools_reg/fit_others/",
#                          package = "magrittr")
batchtools::loadRegistry(file.dir = "r_batchtools_reg/fit_others", 
                         writeable = TRUE)
batchtools::clearRegistry()

# Grid parameters
ncpus <- 1
partition <- "shared"
walltime <- 1 * 3600

# Output dir
dir_output <- "results/fitted/Others/"
dir.create(dir_output, recursive = TRUE, showWarnings = FALSE)
```

```{r setup simulation params}
load("results/fitted/tb_data.RData")
tb_job_others <- tb_data %>% 
  tidyr::expand_grid(
    method = c("metaSPARSim", "DM")
  ) %>% 
  dplyr::mutate(i_job = seq_len(dplyr::n()))
save(tb_job_others, file = paste0(dir_output, "tb_job_others.RData"))
```

```{r simulation task}
one_job <- function(i_job) {
  i_tb_job <- tb_job_others[i_job, ]
  
  load(paste0("data/physeqs/", i_tb_job$dataset, ".RData"))
  x_samples <- get(paste0("physeq_", i_tb_job$dataset, "_bl")) %>% 
  {phyloseq::prune_samples(seq_len(phyloseq::nsamples(.)) %in%
                             i_tb_job$training[[1]],
                           .)} %>% 
    smar::prune_taxaSamples() %>% 
    smar::otu_table2()
  
  # fit methods
  if(i_tb_job$method == "metaSPARSim") {
    require(matrixStats)
    source("R/GMPR.R")
    
    norm_factor <- GMPR(x_samples, intersect.no = 1)
    x_samples_norm <- t(t(x_samples) / norm_factor)
    fit_metaSPARSim <- metaSPARSim::estimate_parameter_from_data(
      raw_data = x_samples,
      norm_data = x_samples_norm,
      conditions = list("group" = colnames(x_samples)),
      perc_not_zeros = 0)
    save(fit_metaSPARSim, file = paste0(dir_output, "fit_", i_job, ".RData"))
  }
  
  if(i_tb_job$method == "DM") {
    fit_DMM <- dirmult::dirmult(t(x_samples),
                                epsilon = 0.01,
                                trace = FALSE)
    save(fit_DMM, file = paste0(dir_output, "fit_", i_job, ".RData"))
  }
}
```

```{r submit jobs}
time_start <- Sys.time()
tb_ids <- batchtools::batchMap(one_job, 
                               i_job = tb_job_others$i_job)
batchtools::batchExport(mget(ls()))
batchtools::submitJobs(ids = tb_ids$job.id,
                       resources =  list(ncpus = ncpus, 
                                         partition = partition, 
                                         walltime = walltime))
print(Sys.time() - time_start)
```
