==> rmds/2_Simulation.Rmd <==
---
title: "Simulate datasets"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE)
library(magrittr)
library(ggplot2)
```

```{r setup2}
# Registry
# batchtools::makeRegistry(file.dir = "r_batchtools_reg/Simulation",
#                          package = "magrittr")
batchtools::loadRegistry(file.dir = "r_batchtools_reg/Simulation", 
                         writeable = TRUE)
batchtools::clearRegistry()

# Grid parameters
ncpus <- 1
partition <- "shared"
walltime <- 5 * 3600

dir_output <- "results/Simulation/"
dir.create(dir_output, recursive = TRUE)
```

```{r job grid}
load("results/fitted/SparseDOSSA2/tb_job_SparseDOSSA2.RData")
tb_fit_SparseDOSSA2 <- tb_job_SparseDOSSA2 %>% 
  dplyr::group_by(i_job) %>% 
  dplyr::mutate(fit = {
    load(paste0("results/fitted/SparseDOSSA2/fit_", i_job, ".RData"))
    list(fit_EM)
  })
tb_fit_SparseDOSSA2 <- tb_fit_SparseDOSSA2 %>%
  dplyr::group_by(i_job) %>%
  dplyr::mutate(logLik = fit[[1]]$EM_fit$fit$logLik,
                logLik_CV =
                  fit[[1]]$EM_fit$logLik_CV[, 1] %>%
                  setdiff(-Inf) %>% 
                  mean(),
                logLik_CV_sd =
                  fit[[1]]$EM_fit$logLik_CV[, 1] %>%
                  setdiff(-Inf) %>% 
                  {sd(.)/sqrt(length(.))}) %>%
  dplyr::ungroup()

tb_fit_SparseDOSSA2 <- tb_fit_SparseDOSSA2 %>%
  dplyr::group_by(dataset, K_split) %>%
  dplyr::arrange(-logLik_CV) %>%
  dplyr::slice(1) %>% 
  dplyr::ungroup()

load("results/fitted/Others/tb_job_others.RData")
tb_fit_others <- tb_job_others %>% 
  dplyr::group_by(i_job) %>% 
  dplyr::mutate(fit = {
    load(paste0("results/fitted/Others/fit_", i_job, ".RData"))
    list(get(paste0("fit_", method)))
  }) %>% 
  dplyr::ungroup()

tb_job <- rbind(
  tb_fit_SparseDOSSA2 %>% 
    dplyr::rename(i_job_fit = i_job),
  tb_fit_others %>% 
    dplyr::rename(i_job_fit = i_job) %>% 
    dplyr::mutate(lambdas = NA,
                  logLik = NA,
                  logLik_CV = NA,
                  logLik_CV_sd = NA)
) %>% 
  tidyr::expand_grid(R = seq(1, 5)) %>% 
  dplyr::mutate(i_job = seq(1, dplyr::n()))
jobs <- tb_job$i_job
save(tb_job, file = paste0(dir_output, "tb_job.RData"))
rm(list = c("tb_job_SparseDOSSA2",
            "tb_job_others",
            "tb_fit_SparseDOSSA2",
            "tb_fit_others",
            "tb_job"))
```

```{r define one job}
one_job <- function(i_job) {
  set.seed(i_job)
  load(paste0(dir_output, "tb_job.RData"))
  i_tb_job <- tb_job[tb_job$i_job == i_job, ]
  
  load(paste0("data/physeqs/", i_tb_job$dataset, ".RData"))
  n_sample <- length(i_tb_job$training[[1]])
  lib_size <- get(paste0("physeq_", i_tb_job$dataset, "_bl")) %>% 
    smar::sample_data2() %>% {.$read_depth[i_tb_job$training[[1]]]}
  fit <- i_tb_job$fit[[1]]
  
  time_start <- Sys.time()
  if(i_tb_job$method == "SparseDOSSA2") {
    sim_SparseDOSSA2 <- SparseDOSSA2::SparseDOSSA2(template = fit,
                                                   n_sample = n_sample,
                                                   new_feature = FALSE, 
                                                   spike_metadata = "none")
    time <- Sys.time() - time_start
    save(sim_SparseDOSSA2, 
         file = paste0(dir_output, "SparseDOSSA2_", i_job, ".RData"))
    sim_data <- vapply(
      seq_len(n_sample),
      function(i_sample) {
        rmultinom(
          n = 1,
          size = lib_size[i_sample],
          prob = sim_SparseDOSSA2$simulated_matrices$rel[, i_sample])
      },
      rep(0, nrow(sim_SparseDOSSA2$simulated_matrices$rel)))
    sim_data <- sim_SparseDOSSA2$simulated_data
    taxa_names <- fit$EM_fit$fit$pi0 %>% names()
  }
    
  if(i_tb_job$method == "metaSPARSim") {
    sim_metaSPARSim <- metaSPARSim::metaSPARSim(fit)
    time <- Sys.time() - time_start
    save(sim_metaSPARSim, file = paste0(dir_output, "metaSPARSim_", i_job, ".RData"))
    sim_data <- sim_metaSPARSim$counts
    taxa_names <- fit$group$intensity %>% names()
  }
  
  if(i_tb_job$method == "DM") {
    sim_DM <- 
      MCMCpack::rdirichlet(n = n_sample,
                           alpha = fit$gamma) %>% 
      t()
    save(sim_DM, file = paste0(dir_output, "DM_", i_job, ".RData"))
    sim_data <- vapply(seq_len(n_sample),
                       function(i_sample) {
                         rmultinom(n = 1,
                                   size = lib_size[i_sample],
                                   prob = sim_DM[, i_sample])
                       },
                       rep(0, nrow(sim_DM)))
    time <- Sys.time() - time_start
        taxa_names <- names(fit$pi)
  }
  
  dimnames(sim_data) <- list(taxa_names,
                             paste0("Sample", seq_len(n_sample)))

  save(sim_data, file = paste0(dir_output, i_job, ".RData"))
  save(time, file = paste0(dir_output, "time_", i_job, ".RData"))
}
```

```{r submit jobs}
tb_ids <- batchtools::batchMap(one_job,
                               i_job = jobs)
batchtools::batchExport(mget(ls()))
batchtools::submitJobs(tb_ids$job.id,
                       resources =  list(ncpus = ncpus, 
                                         partition = partition, 
                                         walltime = walltime))
```
