==> rmds/3_Evaluate.Rmd <==
---
title: "Compare methods through R2"
author: "Siyuan Ma"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE)
library(magrittr)
library(ggplot2)
```

```{r source functions, include=FALSE}
library(magrittr)
library(ggplot2)

dir_output <- "results/Evaluate/"
dir.create(dir_output, recursive = TRUE, showWarnings = FALSE)
```

```{r setup2}
# Registry
# batchtools::makeRegistry(file.dir = "r_batchtools_reg/Evaluation",
#                          package = "magrittr")
batchtools::loadRegistry(file.dir = "r_batchtools_reg/Evaluation", 
                         writeable = TRUE)
batchtools::clearRegistry()

# Grid parameters
ncpus <- 1
partition <- "shared"
walltime <- 3600

dir_output <- "results/Evaluate/"
dir.create(dir_output, recursive = TRUE)
```

```{r job grid}
load("results/Simulation/tb_job.RData")
tb_job <- tb_job %>% 
  dplyr::rename(i_job_sim = i_job) %>% 
  dplyr::select(dataset, K_split, training, method, i_job_sim, R)
tb_job <- tb_job %>% 
  dplyr::group_by(dataset, K_split) %>% 
  dplyr::slice(1) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(method = "Original",
                i_job_sim = NA) %>% 
  rbind(tb_job) %>% 
  dplyr::mutate(i_job = seq(1, dplyr::n()))
save(tb_job, file = paste0(dir_output, "tb_job.RData"))
```

```{r one job}
one_job <- function(i_job) {
  i_tb_job <- tb_job[i_job, ]
  
  load(paste0("data/physeqs/", i_tb_job$dataset, ".RData"))
  
  # Evaluate R2
  original_data <- get(paste0("physeq_", i_tb_job$dataset, "_bl")) %>% 
    smar::otu_table2()
  sim_data_fill <- original_data[, i_tb_job$training[[1]]]
  if(i_tb_job$method != "Original") {
    load(paste0("results/Simulation/", i_tb_job$i_job_sim, ".RData"))
    sim_data_fill <- original_data[, i_tb_job$training[[1]]]
    colnames(sim_data_fill) <- paste0("Sample", seq_len(ncol(sim_data_fill)))
    sim_data_fill[rownames(sim_data), ] <- sim_data
  }
  original_data <- original_data[, setdiff(seq_len(ncol(original_data)),
                                           i_tb_job$training[[1]])]
  
  dataset_combined <- cbind(original_data, sim_data_fill) %>% 
    apply(2, function(x) x / sum(x))
  D <- vegan::vegdist(t(dataset_combined), method = "bray")
  R2 <- vegan::adonis(D ~ sample,
                      data = data.frame(sample = c(rep("original", ncol(original_data)),
                                                   rep("simulation", ncol(sim_data_fill)))),
                      permutations = 2)$aov.tab$R2[1]
  return(R2)
}
```

```{r submit jobs}
tb_ids <- batchtools::batchMap(one_job,
                               i_job = tb_job$i_job)
batchtools::batchExport(mget(ls()))
batchtools::submitJobs(ids = tb_ids$job.id,
                       resources =  list(ncpus = ncpus, 
                                         partition = partition, 
                                         walltime = walltime))
```

```{r gather results}
R2 <- sapply(tb_job$i_job, batchtools::loadResult)
save(R2, file = paste0(dir_output, "R2.RData"))
```
