==> rmds/4.05_analyze_benchmarking.Rmd <==
---
title: "Different benchmarking evaluation metrics.Rmd"
output: html_document
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
library(magrittr)
library(ggplot2)
```

```{r load results}
# generate results
load("results/all_comparison/tb_job.RData")
# batchtools::loadRegistry(file.dir = "r_batchtools_reg/all_comparison", 
#                          writeable = FALSE)
tb_results <- tb_job %>% 
  tidyr::expand_grid(r = seq(1, 20)) %>% 
  # dplyr::filter(i_job %in% batchtools::findDone()$job.id) %>% 
  dplyr::group_by(i_job, r) %>% 
  dplyr::mutate(fit = {
    load(paste0("results/all_comparison/fit_", i_job, "_", r, ".RData"))
    list(fit)
  }) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(Method = factor(method, levels = c("DESeq2", "edgeR", "limmaVOOM", "ANCOM", "MaAsLin2")))

tb_summary_null <- tb_results %>% 
  dplyr::filter(effect == 0) %>% 
  dplyr::group_by(i_job, r) %>% 
  dplyr::mutate(
    n_FP_q = 
      sum(fit[[1]]$qval < 0.05, na.rm = TRUE),
    n_P_q = 
      sum(fit[[1]]$qval < 0.05, na.rm = TRUE),
    n_FP_p = 
      sum(fit[[1]]$pval < 0.05, na.rm = TRUE),
    n_P_p = 
      sum(fit[[1]]$pval < 0.05, na.rm = TRUE)
    ) %>% 
  dplyr::group_by(n, Method, sim_method) %>%
  dplyr::summarise(
    FDR = mean(n_FP_q / (n_P_q + 1e-10)),
    FDR_sd = sd(n_FP_q / (n_P_q + 1e-10)) / sqrt(500),
    FPR = mean(n_FP_p / 330),
    FPR_sd = sd(n_FP_p / 330) / sqrt(500)
  ) %>%
  dplyr::ungroup()

tb_summary_nonNull <- tb_results %>% 
  dplyr::filter(effect > 0) %>% 
  dplyr::group_by(i_job, r) %>% 
  dplyr::mutate(
    n_TP_p = 
      sum(fit[[1]]$pval < 0.05 & stringr::str_detect(fit[[1]]$feature, "TP"), na.rm = TRUE),
    n_FP_p = 
      sum(fit[[1]]$pval < 0.05 & !stringr::str_detect(fit[[1]]$feature, "TP"), na.rm = TRUE),
    n_T = sum(stringr::str_detect(fit[[1]]$feature, "TP")),
    n_TP_q = 
      sum(fit[[1]]$qval < 0.05 & stringr::str_detect(fit[[1]]$feature, "TP"), na.rm = TRUE),
    n_FP_q = 
      sum(fit[[1]]$qval < 0.05 & !stringr::str_detect(fit[[1]]$feature, "TP"), na.rm = TRUE),
    n_P_q = sum(fit[[1]]$qval < 0.05, na.rm = TRUE)
    ) %>% 
  dplyr::group_by(n, Method, sim_method, effect) %>%
  dplyr::summarise(
    Power = mean(n_TP_p / n_T),
    Power_sd = sd(n_TP_p / n_T) / sqrt(500),
    FPR = mean(n_FP_p / 315),
    FPR_sd = sd(n_FP_p / 315) / sqrt(500),
    Recall = mean(n_TP_q / n_T),
    Recall_sd = sd(n_TP_q / n_T) / sqrt(500),
    FDR = mean(n_FP_q / (n_P_q + 1e-10)),
    FDR_sd = sd(n_FP_q / (n_P_q + 1e-10)) / sqrt(500)
  ) %>%
  dplyr::ungroup()

save(tb_summary_null, file = "results/all_comparison/tb_summary_null.RData")
save(tb_summary_nonNull, file = "results/all_comparison/tb_summary_nonNull.RData")
```
# Null case
```{r null case, fig.width = 4, fig.height = 3.5}
p_FDR <- tb_summary_null %>% 
  ggplot(aes(x = factor(n),
             y = FDR,
             color = Method)) +
  geom_point(position = position_dodge(width = 0.25)) +
  geom_errorbar(aes(ymin = FDR - FDR_sd,
                    ymax = FDR + FDR_sd),
                width = 0.25,
                position = position_dodge(width = 0.25)) +
  facet_wrap(~sim_method) +
  theme_bw() +
  theme(legend.position = c(0.01, 0.99),
        legend.justification = c(0, 1),
        legend.background = element_blank(),
        axis.title.x = element_blank()) +
  ggtitle("Null case FDR")
p_FPR <- tb_summary_null %>% 
  ggplot(aes(x = factor(n),
             y = FPR,
             color = Method)) +
  geom_point(position = position_dodge(width = 0.25)) +
  geom_errorbar(aes(ymin = FPR - FPR_sd,
                    ymax = FPR + FPR_sd),
                width = 0.25,
                position = position_dodge(width = 0.25)) +
  facet_wrap(~sim_method) +
  theme_bw() +
  theme(legend.position = c(0.01, 0.99),
        legend.justification = c(0, 1),
        legend.background = element_blank(),
        axis.title.x = element_blank()) +
  ggtitle("Null case FPR")
print(p_FDR)
print(p_FPR)
```

# Spiked-in case

```{r spiked-in, fig.width=8, fig.height=4}
p_FDR <- tb_summary_nonNull %>% 
  ggplot(aes(x = factor(n),
             y = FDR,
             color = Method)) +
  geom_point(position = position_dodge(width = 0.25)) +
  geom_errorbar(aes(ymin = FDR - FDR_sd,
                    ymax = FDR + FDR_sd),
                width = 0.25,
                position = position_dodge(width = 0.25)) +
  facet_wrap(~effect, nrow = 1) +
  theme_bw() +
  theme(legend.position = c(0.01, 0.99),
        legend.justification = c(0, 1),
        legend.background = element_blank(),
        axis.title.x = element_blank()) +
  ggtitle("Non-null case FDR")
p_Recall <- tb_summary_nonNull %>% 
  ggplot(aes(x = factor(n),
             y = Recall,
             color = Method)) +
  geom_point(position = position_dodge(width = 0.25)) +
  geom_errorbar(aes(ymin = Recall - Recall_sd,
                    ymax = Recall + Recall_sd),
                width = 0.25,
                position = position_dodge(width = 0.25)) +
  facet_wrap(~effect, nrow = 1) +
  theme_bw() +
  theme(legend.position = c(0.01, 0.99),
        legend.justification = c(0, 1),
        legend.background = element_blank(),
        axis.title.x = element_blank()) +
  ggtitle("Non-null case Recall")
p_FPR <- tb_summary_nonNull %>% 
  ggplot(aes(x = factor(n),
             y = FPR,
             color = Method)) +
  geom_point(position = position_dodge(width = 0.25)) +
  geom_errorbar(aes(ymin = FPR - FPR_sd,
                    ymax = FPR + FPR_sd),
                width = 0.25,
                position = position_dodge(width = 0.25)) +
  facet_wrap(~effect, nrow = 1) +
  theme_bw() +
  theme(legend.position = c(0.01, 0.99),
        legend.justification = c(0, 1),
        legend.background = element_blank(),
        axis.title.x = element_blank()) +
  ggtitle("Non-null case FPR")
p_Power <- tb_summary_nonNull %>% 
  ggplot(aes(x = factor(n),
             y = Power,
             color = Method)) +
  geom_point(position = position_dodge(width = 0.25)) +
  geom_errorbar(aes(ymin = Power - Power_sd,
                    ymax = Power + Power_sd),
                width = 0.25,
                position = position_dodge(width = 0.25)) +
  facet_wrap(~effect, nrow = 1) +
  theme_bw() +
  theme(legend.position = c(0.01, 0.99),
        legend.justification = c(0, 1),
        legend.background = element_blank(),
        axis.title.x = element_blank()) +
  ggtitle("Non-null case power")
print(p_FDR)
print(p_Recall)
print(p_FPR)
print(p_Power)
```
